v1_0_7

+++ User/dash_draw.c
@@ -84,7 +84,7 @@ int32_t getODValue( VIRTUAL_CHANNEL_t virtualchannel, uint8_t offset_enable)
 
         data16 = GetRPM(INPUT_2);
         float coof2 = (float)getReg16(RPM2_COOF)/RMP_OFFSET;
-        return ( ((u32)(data16 *coof2))*10 );
+        return  (u32)(((float)data16 *coof2)*10 );
       case chKEY  :
           return ((u32) keystate);
       case chDIN1   :
